requirements.txt
streamlit==1.3.0
torch==1.10.2
torchvision==0.11.3
Pillow==9.0.1
matplotlib==3.5.1
numpy==1.22.3
azureml-core==1.39.0
requests==2.27.1

